Move the Ban stuff to it's own directory, and rename the ban collection to Bans.
